fix: msw 상호 참조 에러 수정 (#ATR-58)
