+++ apps/service/app/layout.tsx
@@ -1,14 +1,11 @@
 import type { Metadata } from 'next'
-import { Inter } from 'next/font/google'
-import './globals.css'
 import '@/public/fonts/fonts.css'
+import './globals.css'
 import initMSW from '@/__mocks__'
 
 if (process.env.NODE_ENV !== 'production') {
   initMSW()
 }
-
-const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
   title: 'Create Next App',
@@ -22,7 +19,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body>{children}</body>
     </html>
   )
 }
